ECS System finished (untested)

+++ Void/src/Void/Platform/FMOD/AudioManager/FMODAudioManagerManipulation.cpp
@@ -5,7 +5,7 @@
 namespace Void::Audio {
 	void FMODAudioManager::SetChannelVolume(unsigned int channelID, unsigned int volume)
 	{
-		FMOD_RESULT result = m_ChannelsPool->GetPoolObject(channelID)->FMODChannel->setVolume(volume);
+		FMOD_RESULT result = m_ChannelsPool->GetPoolObject(channelID)->FMODChannel->setVolume((float)volume);
 		VOID_ERROR("Error while setting the volume of channel: '{0}' - {1}", channelID, FMOD_ErrorString(result));
 		m_ChannelsPool->GetPoolObject(channelID)->Volume = volume;
 	}
@@ -27,7 +27,7 @@ namespace Void::Audio {
 	void FMODAudioManager::SetChannelGroupVolume(unsigned int groupID, unsigned int volume)
 	{
 		VOID_ASSERT(groupID < m_ChannelGroups.size(), "ID out of bounds.");
-		FMOD_RESULT result = m_ChannelGroups[groupID]->ChannelGroup->setVolume(volume);
+		FMOD_RESULT result = m_ChannelGroups[groupID]->ChannelGroup->setVolume((float)volume);
 		VOID_ERROR("Error while setting the volume of channelGroup: '{0}' - {1}", groupID, FMOD_ErrorString(result));
 		m_ChannelGroups[groupID]->Volume = volume;
 	}
